refactor: organise tests into module to maximise code re-use (#111)
